Upd: 1. Затянул исправления из ДЗ2 2. Общий index.jsx для компонентов Table 3. Размеры TableCell теперь через классы 4. TableFooter привел к виду общего компонента
