+++ src/components/modules/OrderTable/OrderTable.module.css
@@ -1,0 +1,24 @@
+.checkbox {
+  min-width: 2rem;
+}
+
+.orderNumber {
+  min-width: 6.75rem;
+}
+
+.date,
+.status {
+  min-width: 11rem;
+}
+
+.positionCount {
+  min-width: 6rem;
+}
+
+.amount {
+  min-width: 8rem;
+}
+
+.name {
+  width: 100%;
+}
