+++ src/components/modules/OrderTable/index.jsx
@@ -1,0 +1,3 @@
+import OrderTable from "components/modules/OrderTable/OrderTable";
+
+export default OrderTable;
