+++ src/components/modules/OrderTableBody/index.jsx
@@ -1,0 +1,3 @@
+import OrderTableBody from "./OrderTableBody";
+
+export default OrderTableBody;
