+++ src/components/modules/OrderTableFooter/OrderTableFooter.module.css
@@ -1,0 +1,34 @@
+.leftBlock {
+  display: grid;
+  grid-auto-flow: column;
+  grid-column-gap: 1rem;
+  align-items: center;
+  margin-left: 0.5rem;
+}
+
+.rightBlock {
+  display: grid;
+  grid-auto-flow: column;
+  grid-column-gap: 1rem;
+  align-items: center;
+}
+
+.pages {
+  display: grid;
+  grid-auto-flow: column;
+  grid-column-gap: 0.25rem;
+  align-items: center;
+}
+
+.deletionDropdownWrapper,
+.changeStateDropdownWrapper {
+  left: 0;
+  transform: translateY(-100%);
+  top: -0.25rem;
+}
+
+.pageSelectionDropdownWrapper {
+  right: 0;
+  transform: translateY(-100%);
+  top: -0.25rem;
+}
