+++ src/components/modules/OrderTableFooter/index.jsx
@@ -1,0 +1,3 @@
+import OrderTableFooter from "./OrderTableFooter";
+
+export default OrderTableFooter;
