+++ src/components/modules/OrderTableHeader/index.jsx
@@ -1,0 +1,3 @@
+import OrderTableHeader from "./OrderTableHeader";
+
+export default OrderTableHeader;
