--- src/components/modules/TableContainer/index.jsx
@@ -1,3 +1,0 @@
-import TableContainer from "./TableContainer";
-
-export default TableContainer;
