+++ src/components/pages/OrderPage/OrderPage.jsx
@@ -2,7 +2,7 @@ import ordersMock from "components/mocks/ordersMock100Records";
 import Header from "components/modules/Header";
 import Filter from "components/modules/Filter";
 import { FilterProvider } from "components/context/FilterContext";
-import TableContainer from "components/modules/TableContainer";
+import OrderTable from "components/modules/OrderTable";
 import styles from "./OrderPage.module.css";
 
 function OrderPage() {
@@ -12,7 +12,7 @@ function OrderPage() {
       <FilterProvider>
         <Filter />
       </FilterProvider>
-      <TableContainer orders={ordersMock} />
+      <OrderTable orders={ordersMock} />
     </div>
   );
 }
