--- src/components/shared/Table/TableBody/index.jsx
@@ -1,3 +1,0 @@
-import TableBody from "./TableBody";
-
-export default TableBody;
