--- src/components/shared/Table/TableCell/index.jsx
@@ -1,1 +1,0 @@
-export { default as TableCell, CELL_TYPE } from "./TableCell";
