+++ src/components/shared/Table/TableFooter/TableFooter.module.css
@@ -5,38 +5,3 @@
   background-color: var(--color-day-white-blue, var(--color-night-white-blue));
   color: var(--color-blue);
 }
-
-.leftBlock {
-  display: grid;
-  grid-auto-flow: column;
-  grid-column-gap: 1rem;
-  align-items: center;
-  margin-left: 0.5rem;
-}
-
-.rightBlock {
-  display: grid;
-  grid-auto-flow: column;
-  grid-column-gap: 1rem;
-  align-items: center;
-}
-
-.pages {
-  display: grid;
-  grid-auto-flow: column;
-  grid-column-gap: 0.25rem;
-  align-items: center;
-}
-
-.deletionDropdownWrapper,
-.changeStateDropdownWrapper {
-  left: 0;
-  transform: translateY(-100%);
-  top: -0.25rem;
-}
-
-.pageSelectionDropdownWrapper {
-  right: 0;
-  transform: translateY(-100%);
-  top: -0.25rem;
-}
