--- src/components/shared/Table/TableFooter/index.jsx
@@ -1,3 +1,0 @@
-import TableFooter from "./TableFooter";
-
-export default TableFooter;
