--- src/components/shared/Table/TableHeader/index.jsx
@@ -1,3 +1,0 @@
-import TableHeader from "./TableHeader";
-
-export default TableHeader;
