--- src/components/shared/Table/TableHeaderCell/index.jsx
@@ -1,3 +1,0 @@
-import TableHeaderCell from "./TableHeaderCell";
-
-export default TableHeaderCell;
