--- src/components/shared/Table/TableRow/index.jsx
@@ -1,3 +1,0 @@
-import TableRow from "./TableRow";
-
-export default TableRow;
