+++ src/components/shared/Table/index.jsx
@@ -1,3 +1,7 @@
-import Table from "./Table";
-
-export default Table;
+export { default as Table } from "./Table";
+export { default as TableBody } from "./TableBody/TableBody";
+export { default as TableCell } from "./TableCell/TableCell";
+export { default as TableFooter } from "./TableFooter/TableFooter";
+export { default as TableHeader } from "./TableHeader/TableHeader";
+export { default as TableHeaderCell } from "./TableHeaderCell/TableHeaderCell";
+export { default as TableRow } from "./TableRow/TableRow";
